iio: pressure: bm1390: Add note about driver status
This driver version will be provided as an alpha release to people who
need to urgently get things moving. The raw-reads for pressure and
temperature seem to be somehow working - rest is not.

Add below comment to explain that this is not intended to be the final
release - but is just an early release - improved versions are to be
expected still later 2023.

 * Please note, this driver is in alpha stage. The trigger and buffer work is
 * not completed and is thus disabled. The driver will be further developed
 * and plan is to submit it in upstream Linux kernel when work is finished.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>