iio: pressure: rohm-bm1390: Fix temperature scale
The IIO sysfs ABI mandates the temperature to be in units of
millicelsius (after the scale and offset are applied).. The driver is
errorneously scaling temperatures to kilocelsius.

Fix the scale so that units are correct when scale is applied.

NOTE: This fix is not backwards compatible and thus breaks any existing
user-space applications that have used the incorrect temperature scale.
Still, picking this fix now is better than wondering later why things
fall apart when upstream driver (or any other sensor) is used with the
application tuned to work with this old incorrect version.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Fixes: e2acbee1285d ("iio: pressure: Support ROHM BU1390")