iio: bu27034: Fix lux calculation
The BU27034ANUC has different Lux calculation than the old unsupported
BU27034NUC had. The old calculation used 3 different set of the
coefficients, depending on the D1 / D0 ratio.

The BU27034ANUC has only two different set of coefficients. The third
set of coefficients was removed from the coefficient table when Luc
calculation was changed to work with the new BU27034ANUC, but the
calling code still used the old D1 / D0 ratio logic. This results
slightly incorrect lux values when D1 / D0 is such that the existing
coefficient sets are used. Furthermore, when the D1 / D0 ratio is such
that calling code tries to use the removed set of coefficient, the luc
computation returns -EINVAL and readin returns error to the user.

Fix the calling code to use correct D1 / D0 logic when picking the right
coefficients.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>