kx022a: kx132acr-lbz: Add front matter with permalink
We want to keep the main text in the README.md so the file gets rendered
on the github source repo. When the front matter is added to the
README.md, the processed content on github pages "web view" will create
a README.html - which is no longer visible with the URL pointing to the
folder containing the file. Eg, README.md no longer works as a "index"
file for pages.

We want to add the front matter so we can add page variables like a
sensor name and compatible - so we can use included template text where
we just replace things like sensor name based on the page variables.

Adding a permalink to the index pages which use front matter is a way to
paper over this problem by making the page to appear on URL the
permalink enty instructs.

Add a front matter with permalink to the folder for kx022a pages so they
can later be made to use a generic device-tree example with a compatible
and a sensor name being variables.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>