web: upstream-status-template: Improve information
We shouldn't say we have created a driver when we have added support to
our IC in an existing driver. That's not true, and it does not give the
credit to the original authors whose work we build on. It's better to
say we have added support for our IC, in an existing driver. That's
true, even if adding the support only meant updating the dt-bindings with proper compatible.

Add a flag which can be used to mark drivers which aren't authored by
us, but which are only re-used by us. Change the driver's description
based on this flag.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>